+++ tests/unit/utils/test_chat.py
@@ -57,7 +57,17 @@ class TestConversation:
     @pytest.mark.parametrize("stream", [True, False])
     @patch("builtins.input", side_effect=["foo", "exit"])
     def test_start(self, mock_input, stream):
-        mock_handle_output = Mock()
+        mock = Mock()
+        # create mock function
+        if stream:
+
+            def mock_handle_output(text: str, *, stream: bool = False):  # pyright: ignore[reportRedeclaration]
+                mock(text, stream=stream)
+        else:
+
+            def mock_handle_output(text: str):
+                mock(text)
+
         agent = Agent(conversation_memory=ConversationMemory(), stream=stream)
 
         chat = Chat(agent, intro_text="foo", handle_output=mock_handle_output)
@@ -66,21 +76,56 @@ class TestConversation:
 
         mock_input.assert_has_calls([call(), call()])
         if stream:
-            mock_handle_output.assert_has_calls(
+            mock.assert_has_calls(
                 [
-                    call("foo"),
-                    call("Thinking..."),
+                    call("foo", stream=False),
+                    call("Thinking...", stream=False),
                     call("Assistant: mock output", stream=True),
                     call("\n", stream=True),
-                    call("Exiting..."),
+                    call("Exiting...", stream=False),
                 ]
             )
         else:
-            mock_handle_output.assert_has_calls(
+            mock.assert_has_calls(
                 [
                     call("foo"),
                     call("Thinking..."),
                     call("Assistant: mock output"),
                     call("Exiting..."),
+                ]
+            )
+
+    @pytest.mark.parametrize("stream", [True, False])
+    @patch("builtins.input", side_effect=["foo", "exit"])
+    def test_start_with_handle_output_kwargs(self, mock_input, stream):
+        mock = Mock()
+
+        def mock_handle_output(text: str, **kwargs):
+            mock(text, stream=kwargs.get("stream", False))
+
+        agent = Agent(conversation_memory=ConversationMemory(), stream=stream)
+
+        chat = Chat(agent, intro_text="foo", handle_output=mock_handle_output)
+
+        chat.start()
+
+        mock_input.assert_has_calls([call(), call()])
+        if stream:
+            mock.assert_has_calls(
+                [
+                    call("foo", stream=False),
+                    call("Thinking...", stream=False),
+                    call("Assistant: mock output", stream=True),
+                    call("\n", stream=True),
+                    call("Exiting...", stream=False),
+                ]
+            )
+        else:
+            mock.assert_has_calls(
+                [
+                    call("foo", stream=False),
+                    call("Thinking...", stream=False),
+                    call("Assistant: mock output", stream=False),
+                    call("Exiting...", stream=False),
                 ]
             )