+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - `GriptapeCloudFileManagerDriver.try_save_file` Asset creation utilizes upsert functionality of PUT Asset API.
+- Removed `azure-core` and `azure-storage-blob` dependencies.
+- `GriptapeCloudFileManagerDriver` no longer requires `drivers-file-manager-griptape-cloud` extra.
 
 ## \[0.34.0\] - 2024-10-29
 
@@ -697,6 +699,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Type hint for parameter `azure_ad_token_provider` on Azure OpenAI drivers to `Optional[Callable[[], str]]`.
 - Missing parameters `azure_ad_token` and `azure_ad_token_provider` on the default client for `AzureOpenAiCompletionPromptDriver`.
+- Breaking change in `Chat.handle_output` behavior.
 
 ## \[0.24.2\] - 2024-04-04